requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrdict==2.0.1
backports.zoneinfo==0.2.1
base58==2.1.1
billiard==4.1.0
black==23.3.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptoapis==1.7.0
cytoolz==0.12.1
Django==4.2.2
django-admin==2.0.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-excel-response2==3.0.5
django-filter==23.2
django-jet-reboot==1.3.3
django-six==1.0.5
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-standardized-errors==0.12.5
drf-yasg==1.21.6
ecdsa==0.18.0
eth-abi==2.2.0
eth-account==0.4.0
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
excel-base==1.0.4
Faker==19.2.0
flake8==6.0.0
flower==2.0.0
gunicorn==20.1.0
hexbytes==0.3.1
humanize==4.7.0
idna==3.4
inflection==0.5.1
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.2
jsonrpcclient==4.0.3
kombu==5.3.1
MarkupSafe==2.1.3
mccabe==0.7.0
model-bakery==1.12.0
mypy-extensions==1.0.0
packaging==23.1
parsimonious==0.8.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycodestyle==2.10.0
pycryptodome==3.18.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pypng==0.20220715.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
redgreenunittest==0.1.1
redis==4.6.0
requests==2.31.0
rlp==1.2.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
trx-utils==1.0.4
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.6
xlwt==1.3.0
